+++ presto-main/src/test/java/com/facebook/presto/operator/aggregation/AbstractTestAggregationFunction.java
@@ -12,7 +12,7 @@ public abstract class AbstractTestAggregationFunction
 {
     public abstract BlockCursor getSequenceCursor(int max);
     public abstract AggregationFunction getFunction();
-    public abstract Number getExpectedValue(long positions);
+    public abstract Number getExpectedValue(int positions);
     public abstract Number getActualValue(AggregationFunction function);
 
     @Test
