+++ presto-main/src/test/java/com/facebook/presto/operator/aggregation/TestDoubleSumAggregation.java
@@ -20,14 +20,14 @@ public class TestDoubleSumAggregation
     }
 
     @Override
-    public Double getExpectedValue(long positions)
+    public Double getExpectedValue(int positions)
     {
         if (positions == 0) {
             return null;
         }
 
         double sum = 0;
-        for (long i = 0; i < positions; i++) {
+        for (int i = 0; i < positions; i++) {
             sum += i;
         }
         return sum;
